Add files via upload
apple talk - AppleTalk is a discontinued proprietary suite of networking protocols developed by AppleComputer for their Macintosh computers.

bs4 - Beautiful Soup is a Python package for parsing HTML and XML documents. It creates a parse tree for parsed pages that can be used to extract data from HTML, which is useful for web scraping.

CNN - convolutional neural network - a deep, feed-forward artificial neural network. Remember that feed-forward neural networks are also called multi-layer perceptrons(MLPs), which are the quintessential deep learning models.

crawling - technique to collect data from the web by finding all the URLs for one or multiple domains. Python has several popular web crawling libraries and frameworks. In this article, we will first introduce different crawling strategies and use cases.

hidden API - Dynamic web front-ends often store data in javascript variables and then render it as HTML on demand (like page load or user action). This means the data is not visible on the page directly though it's still there.

scp - scp.py is an open-source python library used to send and receive files between client and server using the paramiko transport. It provides us the programming interface to use the SCP1 protocol, which lets a server and client have multiple conversations over a single TCP connection.

scrapers - Web scraping is the process of collecting and parsing raw data from the Web, and the Python community has come up with some pretty powerful web scraping tools. The Internet hosts perhaps the greatest source of information on the planet.

scrapy - Scrapy is a free and open-source web-crawling framework written in Python and developed in Cambuslang. Originally designed for web scraping, it can also be used to extract data using APIs or as a general-purpose web crawler. It is currently maintained by Zyte, a web-scraping development and services company.

screen time - Screen time is the amount of time spent using a device with a screen such as a smartphone, computer, television, or video game console. 

selenium - library to web scrape. 

sitemaps - The Sitemaps protocol allows a webmaster to inform search engines about URLs on a website that are available for crawling. A Sitemap is an XML file that lists the URLs for a site.

skimage - Scikit-image, or skimage, is an open source Python package designed for image preprocessing.

span - The <span> HTML element is a generic inline container for phrasing content, which does not inherently represent anything.

spider - Spider is a smart point-and-click web scraping tool. With Spider, you can turn websites into organized data, download it as JSON or spreadsheet.

wikipediaapi - using wikipedia information, e.g. wikipedia data scraping to csv.



https://medium.com/towards-data-science/keyword-extraction-with-bert-724efca412ea 
http://2017.compciv.org/guide/topics/python-standard-library/csv.html  
https://medium.com/stanford-cs224w/applying-graph-ml-to-classify-il-licit-bitcoin-transactions-fd32a1ff5dab 
https://medium.com/stanford-cs224w/wikinet-an-experiment-in-recurrent-graph-neural-networks-3f149676fbf3?source=collection_home---4------0----------------------- 
https://medium.com/stanford-cs224w/why-should-i-trust-my-graph-neural-network-4d964052bd85 
https://proceedings.mlr.press/v101/zhao19a/zhao19a.pdf  
https://scrapfly.io/blog/how-to-scrape-hidden-web-data/ 
https://www.codegrepper.com/code-examples/python/split+text+by+sentences+python  
https://jonathansoma.com/lede/foundations-2017/classes/adv-scraping/scraping-bbc/  
https://rpubs.com/shibatau/770520  
https://towardsdatascience.com/how-to-use-selenium-to-web-scrape-with-example-80f9b23a843a 
https://www.thedataincubator.com/blog/2016/04/27/nltk-vs-spacy-natural-language-processing-in-python/  
https://medium.com/stanford-cs224w/self-supervised-learning-for-graphs-963e03b9f809  
https://realpython.com/sentiment-analysis-python/ 
https://www.aaai.org/ocs/index.php/FLAIRS/FLAIRS20/paper/viewFile/18426/17540  
https://neptune.ai/blog/web-scraping-and-knowledge-graphs-machine-learning  
https://www.adamsmith.haus/python/answers/how-to-split-text-into-sentences-in-python 
https://medium.com/stanford-cs224w/predicting-los-angeles-traffic-with-graph-neural-networks-52652bc643b1 
https://medium.com/stanford-cs224w/buy-this-session-based-recommendation-using-sr-gnn-d3415e393722 
https://stackoverflow.com/questions/70225452/beautiful-soup-scraping-bbc-sport-website-i-want-to-ignore-scottish-champions 
https://www.nltk.org 
https://medium.com/stanford-cs224w/graphs-are-all-you-need-generating-multimodal-representations-for-vqa-744a8a1ad448  
https://acawiki.org/Home 
https://medium.com/stanford-cs224w/fraud-detection-on-bitcoin-transaction-graphs-using-graph-convolutional-networks-5fc50a903687 
https://medium.com/stanford-cs224w/learning-mesh-based-flow-simulations-on-graph-networks-44983679cf2d 
https://medium.com/stanford-cs224w/gnns-in-neuroscience-graph-convolutional-networks-for-fmri-analysis-8a2e933bd802 
https://medium.com/stanford-cs224w/incorporating-edge-features-into-graph-neural-networks-for-country-gdp-predictions-1d4dea68337d 
https://medium.com/stanford-cs224w/predicting-drug-drug-interactions-using-graph-neural-networks-6a093cc59a33 
https://medium.com/stanford-cs224w/alexa-queue-that-banger-87c79b0631cf 
https://medium.com/stanford-cs224w/deep-learning-on-3d-meshes-9608a5b33c98 
https://tqdm.github.io 
﻿https://medium.com/stanford-cs224w/knowledge-graph-embeddings-simplistic-and-powerful-representations-ed43a1a73c7c 
https://memgraph.com/blog/best-python-packages-tools-for-knowledge-graphs 
https://neptune.ai/blog/web-scraping-and-knowledge-graphs-machine-learning/amp
https://www.jcchouinard.com/wikipedia-api/ 
https://www.warse.org/IJATCSE/static/pdf/file/ijatcse185932020.pdf 
https://www.analyticsvidhya.com/blog/2019/10/how-to-build-knowledge-graph-text-using-spacy/
https://spacy.io/models/en
https://towardsdatascience.com/knowledge-graph-completion-with-pykeen-and-neo4j-6bca734edf43 
https://www.freecodecamp.org/news/scraping-ecommerce-website-with-python/ 
https://www.linkedin.com/learning/web-scraping-with-python/what-is-web-scraping?autoSkip=true&autoplay=true&contextUrn=urn%3Ali%3AlyndaLearningPath%3A56db2f4b3dd5596be4e4989f&resume=false 
https://www.youtube.com/watch?v=lC6mucyD17k
https://github.com/techwithtim/Beautiful-Soup-Tutorial
https://www.analyticsvidhya.com/blog/2018/09/deep-learning-video-classification-python/ 
https://towardsdatascience.com/generate-any-sport-highlights-using-python-3695c98baead 
https://medium.datadriveninvestor.com/analyzing-video-using-python-opencv-and-numpy-5471cab200c4 
https://pypi.org/project/videoanalytics/ 
https://towardsdatascience.com/data-science-for-e-commerce-with-python-a0a97dd7721d 
https://www.kaggle.com/code/izzuddin8803/notebook702220566a 
https://www.kaggle.com/datasets?search=ecommerce 
https://towardsai.net/p/l/ecommerce-data-analysis-for-sales-strategy-using-python 
https://drive.google.com/drive/folders/1J2xPJn14Koe-yXASUTpk4cjYNb4NuZ5Q